refactor(api): PR2 — extract config domain to api/config.py
Move load_config (5-layer precedence: hardcoded → defaults.json → secrets.json
→ config.json → TRADING_* env vars), save_config (with signal_filters and
kill_switch field-level merge semantics), _deep_merge, _load_json_file,
_strip_secrets, _SECRET_KEYS, SignalFiltersUpdate, ConfigUpdate, plus
/config GET+POST routes from btc_api.py to api/config.py.

api/config.py uses api.deps.verify_api_key (already byte-for-byte equivalent
to btc_api.py:267-275). The legacy verify_api_key in btc_api.py remains
for the still-inline routes (/scan, /webhook/test, etc.) until later PRs.

btc_api.py re-exports load_config, save_config, _strip_secrets, _SECRET_KEYS,
SignalFiltersUpdate, ConfigUpdate, _deep_merge, _load_json_file, get_config,
update_config for legacy callers (scanner_loop, position handlers, existing
tests) until PR7.

Tests updated to also patch api.config.CONFIG_FILE (the authoritative module)
wherever btc_api.CONFIG_FILE was previously patched alone — affects TestLoadConfig
(test_api.py), test_webhook_test_con_url (test_api.py), test_health_persistence,
and test_strategy_kill_switch_v2_calibrator.

Parity test (tests/test_api_config_parity.py) backed by snapshot baseline
covers: GET with auth, GET without auth (401), POST without auth (401),
POST with valid update.

tests/_baselines/.gitignore added to un-ignore config.json in that directory
(global .gitignore matches config.json as a secrets guard; the baseline is
a test fixture, not a secrets file).

1002 tests pass.
