+++ tests/_baselines/.gitignore
@@ -1,0 +1,3 @@
+# Un-ignore config.json in this directory — it is a test fixture (parity baseline),
+# not the real production config.json which contains secrets.
+!config.json
